LayoutTests/fast/css/overscroll-behavior-invalidate-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ OverscrollBehaviorEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that overscroll-behavior is not exposed when the feature is disabled");

shouldBeFalse("'overscrollBehavior' in document.documentElement.style");
shouldBeFalse("'overscroll-behavior' in getComputedStyle(document.documentElement)");
shouldBeFalse("CSS.supports('overscroll-behavior: contain')");
shouldBeFalse("CSS.supports('overscroll-behavior: inherit')");
shouldBeFalse("CSS.supports('overscroll-behavior-x: inherit')");
shouldBeFalse("CSS.supports('overscroll-behavior-y: inherit')");

</script>
</body>
</html>
